+++ src/lib/pages/account-details/data.ts
@@ -35,9 +35,11 @@ import {
   addrToValoper,
   addTokenWithValue,
   coinToTokenWithValue,
+  totalValueTokenWithValue,
+  compareTokenWithValues,
 } from "lib/utils";
 
-import type { UserDelegationsData } from "./type";
+import type { DenomInfo, UserDelegationsData } from "./types";
 
 interface AccountContracts {
   contracts: Option<ContractInfo[]>;
@@ -57,28 +59,26 @@ interface AccountAssetInfos {
   error: Error;
 }
 
-export interface StakingParams extends RawStakingParams {
-  symbol: Option<string>;
-  logo: Option<string>;
-  precision: Option<number>;
+export interface StakingParams extends Omit<RawStakingParams, "bondDenoms"> {
+  bondDenoms: DenomInfo[];
 }
 
 export interface Delegation {
   validator: ValidatorInfo;
-  token: TokenWithValue;
+  balances: TokenWithValue[];
 }
 
 export interface Unbonding {
   validator: ValidatorInfo;
   completionTime: Date;
-  token: TokenWithValue;
+  balances: TokenWithValue[];
 }
 
 export interface Redelegation {
   srcValidator: ValidatorInfo;
   dstValidator: ValidatorInfo;
   completionTime: Date;
-  token: TokenWithValue;
+  balances: TokenWithValue[];
 }
 
 export const useAccountContracts = (
@@ -262,13 +262,15 @@ export const useUserDelegationInfos = (walletAddress: HumanAddr) => {
   };
 
   if (rawStakingParams && assetInfos && validators) {
-    const stakingParams = {
+    data.stakingParams = {
       ...rawStakingParams,
-      symbol: assetInfos[rawStakingParams.bondDenom]?.symbol,
-      logo: assetInfos[rawStakingParams.bondDenom]?.logo,
-      precision: assetInfos[rawStakingParams.bondDenom]?.precision,
+      bondDenoms: rawStakingParams.bondDenoms.map((denom) => ({
+        denom,
+        symbol: assetInfos[denom]?.symbol,
+        logo: assetInfos[denom]?.logo,
+        precision: assetInfos[denom]?.precision,
+      })),
     };
-    data.stakingParams = stakingParams;
 
     data.isValidator = Object.keys(validators).includes(
       addrToValoper(walletAddress)
@@ -278,20 +280,25 @@ export const useUserDelegationInfos = (walletAddress: HumanAddr) => {
       validator: {
         validatorAddress: raw.validatorAddress,
         moniker: validators[raw.validatorAddress]?.moniker,
-        identity: raw.identity,
+        identity: validators[raw.validatorAddress]?.identity,
       },
-      token: coinToTokenWithValue(raw.denom, raw.amount, assetInfos[raw.denom]),
+      balances: raw.balance
+        .map((coin) =>
+          coinToTokenWithValue(coin.denom, coin.amount, assetInfos[coin.denom])
+        )
+        .sort(compareTokenWithValues),
     }));
     data.totalDelegations = data.delegations?.reduce<
       Record<string, TokenWithValue>
     >(
-      (total, delegation) => ({
-        ...total,
-        [delegation.token.denom]: addTokenWithValue(
-          total[delegation.token.denom],
-          delegation.token
+      (total, delegation) =>
+        delegation.balances.reduce(
+          (acc, balance) => ({
+            ...acc,
+            [balance.denom]: addTokenWithValue(acc[balance.denom], balance),
+          }),
+          total
         ),
-      }),
       {}
     );
 
@@ -299,34 +306,41 @@ export const useUserDelegationInfos = (walletAddress: HumanAddr) => {
       validator: {
         validatorAddress: raw.validatorAddress,
         moniker: validators[raw.validatorAddress]?.moniker,
-        identity: raw.identity,
+        identity: validators[raw.validatorAddress]?.identity,
       },
       completionTime: raw.completionTime,
-      token: coinToTokenWithValue(
-        rawStakingParams.bondDenom,
-        raw.amount,
-        assetInfos[rawStakingParams.bondDenom]
-      ),
+      balances: raw.balance
+        .map((coin) =>
+          coinToTokenWithValue(coin.denom, coin.amount, assetInfos[coin.denom])
+        )
+        .sort(compareTokenWithValues),
     }));
     data.totalUnbondings = data.unbondings?.reduce<
       Record<string, TokenWithValue>
     >(
-      (total, unbonding) => ({
-        ...total,
-        [unbonding.token.denom]: addTokenWithValue(
-          total[unbonding.token.denom],
-          unbonding.token
+      (total, unbonding) =>
+        unbonding.balances.reduce(
+          (acc, balance) => ({
+            ...acc,
+            [balance.denom]: addTokenWithValue(acc[balance.denom], balance),
+          }),
+          total
         ),
-      }),
       {}
     );
 
     data.rewards = rawRewards?.rewards.reduce<Record<string, TokenWithValue[]>>(
       (prev, raw) => ({
         ...prev,
-        [raw.validatorAddress]: raw.reward.map<TokenWithValue>((coin) =>
-          coinToTokenWithValue(coin.denom, coin.amount, assetInfos[coin.denom])
-        ),
+        [raw.validatorAddress]: raw.reward
+          .map<TokenWithValue>((coin) =>
+            coinToTokenWithValue(
+              coin.denom,
+              coin.amount,
+              assetInfos[coin.denom]
+            )
+          )
+          .sort(compareTokenWithValues),
       }),
       {}
     );
@@ -348,19 +362,19 @@ export const useUserDelegationInfos = (walletAddress: HumanAddr) => {
       srcValidator: {
         validatorAddress: raw.srcValidatorAddress,
         moniker: validators[raw.srcValidatorAddress]?.moniker,
-        identity: raw.srcIdentity,
+        identity: validators[raw.srcValidatorAddress]?.identity,
       },
       dstValidator: {
         validatorAddress: raw.dstValidatorAddress,
         moniker: validators[raw.dstValidatorAddress]?.moniker,
-        identity: raw.dstIdentity,
+        identity: validators[raw.dstValidatorAddress]?.identity,
       },
       completionTime: raw.completionTime,
-      token: coinToTokenWithValue(
-        stakingParams.bondDenom,
-        raw.amount,
-        assetInfos[stakingParams.bondDenom]
-      ),
+      balances: raw.balance
+        .map((coin) =>
+          coinToTokenWithValue(coin.denom, coin.amount, assetInfos[coin.denom])
+        )
+        .sort(compareTokenWithValues),
     }));
 
     data.totalCommission = rawCommission?.commission.reduce<
@@ -416,11 +430,9 @@ export const useAccountTotalValue = (walletAddress: HumanAddr) => {
     : defaultValue;
   return {
     totalAccountValue: totalAssetValue
-      .add(totalBonded[stakingParams.bondDenom]?.value ?? defaultValue)
-      .add(totalRewards[stakingParams.bondDenom]?.value ?? defaultValue)
-      .add(
-        totalCommission[stakingParams.bondDenom]?.value ?? defaultValue
-      ) as USD<Big>,
+      .add(totalValueTokenWithValue(totalBonded, defaultValue))
+      .add(totalValueTokenWithValue(totalRewards, defaultValue))
+      .add(totalValueTokenWithValue(totalCommission, defaultValue)) as USD<Big>,
     isLoading: false,
   };
 };