+++ src/lib/types/index.ts
@@ -15,6 +15,7 @@ export * from "./projects";
 export * from "./proposal";
 export * from "./rpc";
 export * from "./tx";
+export * from "./staking";
 export * from "./upload";
 export * from "./validator";
 export * from "./wallet";